Refactor the whole thing and make it more meta.
